requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-domain==3.14.9
ansi2html==1.8.0
anyio==3.6.2
astroid==2.4.2
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
atlassian-python-api==3.35.0
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.9.3
bidict==0.22.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
Deprecated==1.2.13
dnspython==2.3.0
EdgeGPT==0.0.58.2
eventlet==0.33.3
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3
Flask-uWSGI-WebSocket==0.6.1
frozenlist==1.3.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
idna==2.10
importlib-metadata==6.1.0
install==1.3.5
isort==5.6.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.6.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==0.27.2
outcome==1.2.0
packaging==23.0
paramiko==3.1.0
pathspec==0.11.1
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
Pillow==9.4.0
platformdirs==3.2.0
pycparser==2.21
pycryptodome==3.17
pygame==2.1.3
Pygments==2.14.0
pylint==2.6.0
PyNaCl==1.5.0
PySocks==1.7.1
pyte==0.8.1
pytesseract==0.3.10
python-engineio==4.4.0
python-socketio==5.7.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.2
selenium==4.8.2
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.0.1
toml==0.10.1
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.14
uWSGI==2.0.21
vultr-python==0.1.5
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.12.1
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
